src/glare/Manuscript_Code/glare/requirements.txt
# python=3.9
numpy==1.23.0
pandas==1.5.3
tqdm==4.64.1
scikit-learn>=1.2.1
scipy>=1.10.0
umap-learn==0.5.6
hdbscan==0.8.29
torch>=1.14.0
xgboost==1.7.6
matplotlib>=3.6.3